refactor(sync): extract transport and release_discovery leaves (P2.A)
Move the HTTP transport (mirrors/headers/cascading fetch) and release
discovery (list/latest-by-prefix/asset_url/check_latest) leaves out of
data/sync into a new top-level sync/ tier -- the only layer permitted to
issue HTTP. data/sync re-exports the moved symbols (incl. ReleaseSpec and
the transport singleton) so all existing prts_mcp.data.sync.* / ./sync.js
import paths and the 25 test patch sites keep resolving during the
P2.A->P2.B migration; the release/archive/pair state machine stays in
data/sync until P2.B.

ReleaseSpec moves with discovery (check_latest_release's param type,
self-contained -- no sync->data back-edge); RepoSpec stays (state-machine
only). check_latest_release log lines now originate from
prts_mcp.sync.release_discovery.

Repoint four test_sync_release patches that feed check_latest_release's
internal _get_cascading call to also patch release_discovery._get_cascading
(shared mock): the move shifted that lookup out of the data.sync namespace
(two of the four were silently passing for the wrong reason before).

PY 369 passed / TS typecheck+build clean, 283 passed / check-runtime --full 0 failures.
