+++ python/tests/test_sync_release.py
@@ -135,6 +135,9 @@ class TestSyncRelease:
         with patch(
             "prts_mcp.data.sync._get_cascading",
             side_effect=[release, asset, manifest],
+        ) as cascading, patch(
+            "prts_mcp.sync.release_discovery._get_cascading",
+            cascading,
         ):
             result = sync_release(spec, force_check=True)
         assert result.status == "updated"
@@ -161,6 +164,9 @@ class TestSyncRelease:
         with patch(
             "prts_mcp.data.sync._get_cascading",
             side_effect=[release, asset, manifest],
+        ) as cascading, patch(
+            "prts_mcp.sync.release_discovery._get_cascading",
+            cascading,
         ):
             result = sync_release(spec, force_check=True)
         assert result.status == "offline_fallback"
@@ -179,6 +185,9 @@ class TestSyncRelease:
         with patch(
             "prts_mcp.data.sync._get_cascading",
             side_effect=[release, asset, _AssetNotFoundError("HTTP 404")],
+        ) as cascading, patch(
+            "prts_mcp.sync.release_discovery._get_cascading",
+            cascading,
         ):
             result = sync_release(spec, force_check=True)
         assert result.status == "updated"
@@ -203,6 +212,9 @@ class TestSyncRelease:
                 _mock_asset_response(b"new"),
                 manifest,
             ],
+        ) as cascading, patch(
+            "prts_mcp.sync.release_discovery._get_cascading",
+            cascading,
         ):
             result = sync_release(spec, force_check=True)
         assert result.status == "offline_fallback"
