+++ ts/src/data/sync.ts
@@ -25,18 +25,17 @@ import {
 import { basename, dirname, isAbsolute, join, relative, resolve, sep } from "node:path";
 import { setTimeout as delay } from "node:timers/promises";
 import AdmZip from "adm-zip";
-import type { Dispatcher } from "undici";
+import {
+  AssetNotFoundError,
+  fetchCascading,
+  githubHeaders,
+  parseMirrors,
+  type FetchResponse,
+} from "../sync/transport.js";
 
-const NATIVE_FETCH = globalThis.fetch;
-let envProxyAgent: Dispatcher | undefined;
-
-/** The response surface shared by the global and Undici fetch runtimes. */
-interface FetchResponse {
-  readonly ok: boolean;
-  readonly status: number;
-  json(): Promise<unknown>;
-  arrayBuffer(): Promise<ArrayBuffer>;
-}
+// Re-exported so existing `./sync.js` consumers (imagesSync, sync.test) keep
+// resolving these during the P2.A→P2.B migration.
+export { AssetNotFoundError, fetchCascading, githubHeaders };
 
 /** Versioned contract shared with the arknights-data-pipeline manifest. */
 export const DATA_CONTRACT_VERSION = "prts-mcp-data/v1";
@@ -55,8 +54,6 @@ export const GAMEDATA_FILES: readonly string[] = [
   "zh_CN/gamedata/excel/zone_table.json",
   "zh_CN/gamedata/excel/item_table.json",
 ];
-
-const GITHUB_UA = "PRTS-MCP-Bot/0.1 (Arknights fan-creation helper)";
 
 /** Skip the upstream SHA check if cached data is fresher than this (seconds). */
 const CACHE_TTL_SECONDS = 3600;
@@ -108,115 +105,8 @@ export interface SyncResult {
 // Internal helpers
 // ---------------------------------------------------------------------------
 
-export function githubHeaders(): Record<string, string> {
-  const headers: Record<string, string> = { "User-Agent": GITHUB_UA };
-  const token = process.env["GITHUB_TOKEN"];
-  if (token) headers["Authorization"] = `Bearer ${token}`;
-  return headers;
-}
-
-/**
- * Parse GITHUB_MIRRORS env var into a list of proxy base URLs (trailing slash stripped).
- *
- * Unset / empty  → [] (direct only, no cascade)
- * "https://ghproxy.net"              → ["https://ghproxy.net"]
- * "https://a.example,https://b.example" → ["https://a.example", "https://b.example"]
- *
- * Mirror URL format (ghproxy-style): <mirror>/<original_url>
- * e.g. "https://ghproxy.net/https://raw.githubusercontent.com/..."
- */
-function parseMirrors(): string[] {
-  return (process.env["GITHUB_MIRRORS"] ?? "")
-    .split(",")
-    .map((s) => s.trim().replace(/\/$/, ""))
-    .filter(Boolean);
-}
-
-/** Return [url, mirroredUrl1, mirroredUrl2, ...] */
-function urlCandidates(url: string): string[] {
-  return [url, ...parseMirrors().map((m) => `${m}/${url}`)];
-}
-
-async function fetchWithRuntimeProxy(
-  url: string,
-  options: Omit<RequestInit, "signal">,
-  signal: AbortSignal,
-): Promise<FetchResponse> {
-  // Preserve test/in-process fetch overrides and Bun's native proxy support.
-  if (
-    globalThis.fetch !== NATIVE_FETCH
-    || process.versions.bun
-    || !(
-      process.env["HTTP_PROXY"] || process.env["HTTPS_PROXY"]
-      || process.env["http_proxy"] || process.env["https_proxy"]
-    )
-  ) {
-    return globalThis.fetch(url, { ...options, signal });
-  }
-  const { EnvHttpProxyAgent, fetch: undiciFetch } = await import("undici");
-  if (envProxyAgent === undefined) envProxyAgent = new EnvHttpProxyAgent();
-  const response = await undiciFetch(url, {
-    ...options,
-    signal,
-    dispatcher: envProxyAgent,
-  } as Parameters<typeof undiciFetch>[1]);
-  return response;
-}
-
-/**
- * fetch() wrapper that cascades through URL candidates on failure.
- *
- * - A fresh AbortSignal.timeout is created per attempt so an earlier
- *   timeout does not consume the budget for later candidates.
- * - HTTP 4xx from the direct URL propagates immediately (resource is
- *   genuinely missing — mirrors won't help).
- * - Network error or HTTP 5xx from any candidate → try the next one.
- */
-export async function fetchCascading(
-  url: string,
-  options: Omit<RequestInit, "signal">,
-  timeoutMs: number,
-): Promise<FetchResponse> {
-  const candidates = urlCandidates(url);
-  let lastErr: unknown = new Error("All URL candidates failed");
-  for (let i = 0; i < candidates.length; i++) {
-    try {
-      const res = await fetchWithRuntimeProxy(
-        candidates[i], options, AbortSignal.timeout(timeoutMs),
-      );
-      if (res.ok) return res;
-      // Direct 404 → asset confirmed absent; record the typed error and stop
-      // without trying mirrors. lastErr + break (not throw) so the typed
-      // error survives to the caller even with GITHUB_MIRRORS (#100).
-      if (i === 0 && res.status === 404) {
-        lastErr = new AssetNotFoundError(`HTTP 404: ${candidates[i]}`);
-        break;
-      }
-      lastErr = new Error(`HTTP ${res.status}`);
-      // Other direct 4xx → the resource does not exist; mirrors cannot help.
-      if (i === 0 && res.status >= 400 && res.status < 500) break;
-    } catch (err) {
-      lastErr = err;
-    }
-  }
-  throw lastErr;
-}
-
 function errorMessage(err: unknown): string {
   return err instanceof Error ? err.message : String(err);
-}
-
-/**
- * The direct release URL returned 404 — the asset genuinely does not exist.
- * Distinguished from a mirror 404 (mirror lacks the asset; the release may
- * still exist upstream) so manifest verification skips only on a confirmed
- * upstream 404 and stays fail-closed otherwise (#100).
- */
-export class AssetNotFoundError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "AssetNotFoundError";
-  }
 }
 
 function cacheIsFresh(cache: CacheMeta): boolean {
@@ -243,87 +133,23 @@ function releaseZipError(spec: ReleaseSpec): string | null {
 // Release-based sync (for zh_CN.zip from GitHub Releases)
 // ---------------------------------------------------------------------------
 
-const TAG_PREFIX = "data-";
-
-// ---------------------------------------------------------------------------
-// Release discovery (tag-prefix filtered)
-//
-// The arknights-data-pipeline repo hosts both ``data-*`` and ``images-*``
-// GitHub Releases.  ``/releases/latest`` may point at an ``images-*`` release
-// if GitHub auto-promotes it, so data sync must filter by tag prefix instead.
-// (imagesSync reuses these helpers for its own prefix filtering.)
-// ---------------------------------------------------------------------------
-
-/** Minimal shape of a GitHub Release object (only the fields we read). */
-export interface GithubRelease {
-  tag_name?: unknown;
-  created_at?: unknown;
-  assets?: unknown;
-  [key: string]: unknown;
-}
-
-/** List all non-draft releases. Returns null on any network/API failure. */
-export async function listReleases(
-  owner: string,
-  repo: string,
-  timeoutMs = 10_000,
-): Promise<GithubRelease[] | null> {
-  const url = `https://api.github.com/repos/${owner}/${repo}/releases?per_page=100`;
-  try {
-    const res = await fetchCascading(url, { headers: githubHeaders() }, timeoutMs);
-    const data = await res.json();
-    return Array.isArray(data) ? (data as GithubRelease[]) : null;
-  } catch {
-    return null;
-  }
-}
-
-/** Pick the newest release whose tag starts with *prefix*, sorting by created_at desc. */
-export function latestReleaseByPrefix(
-  releases: GithubRelease[],
-  prefix: string,
-  opts: { excludePrefix?: string } = {},
-): GithubRelease | null {
-  const candidates: GithubRelease[] = [];
-  for (const release of releases) {
-    const tag = release["tag_name"];
-    if (typeof tag !== "string" || !tag.startsWith(prefix)) continue;
-    if (opts.excludePrefix && tag.startsWith(opts.excludePrefix)) continue;
-    candidates.push(release);
-  }
-  if (candidates.length === 0) return null;
-  candidates.sort((a, b) =>
-    String(b["created_at"] ?? "").localeCompare(String(a["created_at"] ?? "")),
-  );
-  return candidates[0];
-}
-
-/** Extract the browser_download_url for *assetName* from a release object. */
-export function assetUrl(release: GithubRelease, assetName: string): string | null {
-  const assets = release["assets"];
-  if (!Array.isArray(assets)) return null;
-  for (const a of assets) {
-    if (typeof a === "object" && a !== null && (a as Record<string, unknown>)["name"] === assetName) {
-      const url = (a as Record<string, unknown>)["browser_download_url"];
-      if (typeof url === "string") return url;
-    }
-  }
-  return null;
-}
-
-/** Describes a GitHub Release asset to download as a local zip. */
-export interface ReleaseSpec {
-  owner: string;
-  repo: string;
-  /** Asset filename in the release, e.g. "zh_CN.zip". */
-  assetName: string;
-  /** Absolute destination path for the downloaded zip. */
-  localZip: string;
-  /** Optional validator returning missing or invalid zip entries. */
-  validateZip?: (zipPath: string) => string[];
-  /** Verify the optional factory manifest when the release provides it. */
-  verifyManifest?: boolean;
-}
+// Release discovery (tag-prefix-filtered list / latest / asset_url /
+// checkLatestRelease) + the ReleaseSpec type live in the
+// sync/releaseDiscovery tier. Re-imported here so the state machine below
+// and existing `./sync.js` consumers keep resolving during P2.A→P2.B.
+import {
+  type ReleaseSpec,
+  TAG_PREFIX,
+  checkLatestRelease,
+} from "../sync/releaseDiscovery.js";
+export {
+  type GithubRelease,
+  type ReleaseSpec,
+  assetUrl,
+  checkLatestRelease,
+  latestReleaseByPrefix,
+  listReleases,
+} from "../sync/releaseDiscovery.js";
 
 /** Describes a GitHub Release zip asset that should be extracted locally. */
 export interface ReleaseArchiveSpec {
@@ -390,29 +216,6 @@ async function saveReleaseMeta(
     files: meta.files,
   }, null, 2), "utf-8");
   await rename(tmp, p);
-}
-
-/**
- * Return the latest ``data-*`` release tag and asset download URL.
- *
- * Uses the releases list API with tag-prefix filtering instead of
- * ``/releases/latest``, because the data-pipeline repo also hosts
- * ``images-*`` releases that may be promoted to "Latest" on GitHub.
- * Returns null on network failure or when no matching release/asset is found.
- */
-export async function checkLatestRelease(
-  spec: ReleaseSpec,
-  timeoutMs = 10_000
-): Promise<{ tag: string; url: string } | null> {
-  const releases = await listReleases(spec.owner, spec.repo, timeoutMs);
-  if (releases === null) return null;
-  const latest = latestReleaseByPrefix(releases, TAG_PREFIX);
-  if (!latest) return null;
-  const tag = latest["tag_name"];
-  if (typeof tag !== "string") return null;
-  const url = assetUrl(latest, spec.assetName);
-  if (!url) return null;
-  return { tag, url };
 }
 
 /**
